+++ src/xml2_node.cpp
@@ -37,8 +37,9 @@ void node_set_name(XPtrNode node, std::string value) {
 CharacterVector node_text(XPtrNode node, bool trim) {
   std::string text = Xml2String(xmlNodeGetContent(node.checked_get())).asStdString();
 
-  if (trim)
-    boost::algorithm::trim(text);
+  if (trim) {
+    boost::algorithm::trim_if(text, (! boost::algorithm::is_print()) || boost::algorithm::is_space());
+  }
 
   return asCharacterVector(text.c_str());
 }