+++ tests/testthat/test-xml_text.R
@@ -25,10 +25,10 @@ test_that("xml_text works properly with xml_nodeset objects", {
     c("This is some text. This is some nested text.", "This is some nested text."))
 })
 
-test_that("xml_text trims whitespace if requested", {
-  x <- read_xml("<p>   Some text    </p>")
+test_that("xml_text trims whitespace if requested, including non-breaking spaces", {
+  x <- read_html("<p>   Some text    &nbsp;</p>")
   expect_identical(xml_text(x),
-    "   Some text    ")
+    "   Some text     ")
 
   expect_identical(xml_text(x, trim = TRUE),
     "Some text")